requirements.txt
tensorflow
streamlit
numpy
Pillow
matplotlib
kagglehub
sklearn